refactor: move calculate_overdue_days function to shared_utils and clean up imports
